:bug: Make sure that paginators are aliased to `pagination.IStaticPage` etc. as the mappers were not enough (#43)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Proprietary
-->
### Description

<!--
Please add any detail or context that would be useful to a reviewer.
-->

Make sure that paginators are aliased to `pagination.IStaticPage` etc.
as the mappers were not enough. It was impossible to cast between the
mapped struct and the various structs we want to map between due to type
mismatches between `client.IStaticPage` and `pagination.IStaticPage`.
This is the simplest solution.

### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [x]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [ ] Additional tests are not required for this change (e.g.
documentation update).

---------

Co-authored-by: joshjennings98 <joshjennings98@users.noreply.github.com>
Co-authored-by: Adrien CABARBAYE <adrien.cabarbaye@arm.com>