✨ Extend the code generator for JobItems (#75)
### Description
This change refactors the generator to be able to handle multiple types
of generation, and extends it with a new type for JobItems, which are
schemas that have the property `x-job: true`.

The code generated for the the JobItems implements missing methods that
would make these JobItem types satisfy the `IAsynchronousJob` interface
defined
[here](https://github.com/ARM-software/embedded-development-services-client-utils/blob/main/utils/job/interfaces.go).

- [X]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [ ] Additional tests are not required for this change (e.g.
documentation update).

---------

Co-authored-by: aorabdel <aorabdel@users.noreply.github.com>