✨ Add IModel generation for Items that are not in Collections (#80)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Proprietary
-->
### Description
The only Items that get extended to implement IModel are Items inside
Collections, which is not the exclusive list of Items we want to extend,
some `x-job` Items are defined but not in a collection.
This changes generates `IModel` methods to job items that are not in
Collections.


### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [ ]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [X] Additional tests are not required for this change (e.g.
documentation update).