📝 Remove code climate references (#92)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Proprietary
-->
### Description

This removes code climate references in files as we have now migrated to
Qlty Cloud.

_NB: The work to integrate Qlty Cloud changes will be a separate PR as
it's dependent on the existence of a repo/organisation level secret for
authentication with Qlty._

### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [ ]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [X] Additional tests are not required for this change (e.g.
documentation update).

---------

Co-authored-by: Adrien CABARBAYE <adrien.cabarbaye@arm.com>
Co-authored-by: acabarbaye <29707928+acabarbaye@users.noreply.github.com>
Co-authored-by: acabarbaye <acabarbaye@users.noreply.github.com>
Co-authored-by: Kem-Gov <14895121+Kem-Gov@users.noreply.github.com>
Co-authored-by: Kem-Gov <Kem-Gov@users.noreply.github.com>