🐛 Extend generator for `x-no-pagination` extension (#96)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Proprietary
-->
### Description

This change ensures the generator only generates pagination functions on
collections if the `x-no-pagination` extension flag is false or
non-existent.

### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [X]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [ ] Additional tests are not required for this change (e.g.
documentation update).

---------

Co-authored-by: Kem-Gov <14895121+Kem-Gov@users.noreply.github.com>
Co-authored-by: Kem-Gov <Kem-Gov@users.noreply.github.com>